arc: Add plugin to create a single jar of the arc tools
Using the sbt command `sbt assembly` this allows us to create a single
self-contained jar of the entire arc codebase. This greatly speeds up
the startup time of the AST dumper as it can be run without involving
sbt. The startup time will be significant when we are adding further
mlir tests.