scala: Stop corrupting large unsigned ints
Unsigned ints should never be sign-extended, as they are sign-less.

An unsigned 32-bit integer with the high bit set cannot be stored in a
Scala Int as it will be coerced into a negative Long by the AST
constructor. The same applies to unsigned 64-bit integers but here we
have to pass the value to the AST constructor as a BigInt.