Remove crate and crate_end operations from the Rust dialect
Remove the crate and the crate_end operations from the rust dialect as
we can get the same functionality by just using the module name as the
name of the resulting crate. Additionally this simplifies lowering
from the Arc dialect to the Rust dialect as we don't have to collect
and insert functions inside a crate operation.