arc-to-rust: Add support for including external Rust code
Using the command line flag `-extra-rust-trailer <file>` together with
`-crate` to `arc-mlir` will insert `<file>` at the end of the
generated rust module. The intended use is to provide a way to insert
testing code in the output and with this change `arc-mlir` will no
longer produce the dummy test module.