arc-to-rust: Implement support for Arc structs in the backend
Arc structs are mapped to Rust structs. This is done by introducing a
new Rust dialect type to represent the structures and extending the
RustPrinterStream to handle the new type. When Rust dialect structs
are instantiated they are uniqued and named. The RustPrinterStream
will ensure that a struct is always defined before the type is used.