rust: Turn Rust dialect tuples into their own type
This is necessary in order to make nested structs and tuples work
properly. Previously tuples were represented as a RustType, but as the
internal representation is just a string describing the type there was
no way to ensure that contained types were emitted correctly.

This patch makes tuples into their own type which, just like structs,
has a vector containing their field types. It also adds the mutually
recursive methods RustStructTypeStorage::emitNestedTypedefs() and
RustTupleTypeStorage::emitNestedTypedefs() which ensure that
definitions for all structs are emitted before they are used.