rust: Fix bug in output of float constants
The hexf crate is used to allow for exact hexadecimal floating point
constants, unfortunately APFloat::convertToHexString() which is used
to convert floats to their textual representation omits the '.' if the
mantissa is 0 which hexf doesn't like. To work around this, simply
scan for a '.' before the 'p', and if it isn't found, patch the
string.