Fix bug when feeding arc-script with source directly
This fixes a small bug which occurred when running arc-script as a
library which takes source directly from a string instead of reading it
from a file.

All items at the top-level should now be children of `crate`, e.g.,
`crate::my_function`.

Prior, the compiler would think items were it was stored in a fake-file
named `main.arc`, which also happened to result in a syntax error by
giving items the path `crate::main.arc::my_function`. This is now fixed.