Update macros and builder interfaces
This adds a new interface for compiling all arc-script files in a crate
with file-name "main.arc".

```rust
// build/main.rs
fn main() {
    arc_script::Builder::process_root()
}
```

It can be used inside build-scripts to pre-build arc-scripts before
compiling the rest of the crate. Scripts can then be included in
Rust-code with this code:

```rust
// src/main.rs
mod main { }
```

This is important because it allows build-artefacts to be cached. The
other previous and still supported way of compiling is to directly do:

```rust
// src/main.rs
mod main { }
```

The `main.arc` should then always be recompiled when the Rust source
changes. Later it might also be good to support something like:

```rust
// build.rs
fn main() {
    arc_script::Builder::new()
      .precompile("src/main.arc");
}
```

So not all "main.arc" files must be compiled.