CI: Check that there are no broken symlinks in the repo
Arc-mlir uses symlinks under arc-mlir/src/tests/arc-script/ to
gradually increase the number of tests that are run. Typically a
symlink is added when an arc-script test is successfully handled by
arc-mlir.

Unfortunately this makes it easy to update files in arc-script/ and
forget to update the links. Therefore add a CI check to ensure that a
PR does not introduce broken symlinks.