Add #[arcorn::rewrite] proc macro for rewriting enums
This adds a clever proc macro which can be used like:

```
enum Foo {
    Bar(i32),
    Baz(Qux)
}

struct Qux {
    x: i32,
    y: i32
}
```

It will rewrite the items into a format which is compatible with Arcon
and Protobuf (prost). The transformed enums can afterwards be accessed
through the `arcorn::{is, unwrap, enwrap}` macros. See their doc-tests for
more documentation.