Fix bugs
Fixes a bug which prevented operator constructors from compiling if
those constructors borrowed their environment. For example, a constructor
which took a function pointer as input. This was solved by `move`ing the
environment so the constructor could take ownership of it.

Also, this solves a couple of bugs in the snapshot tests which I did not
know about. Apparently some test directories were never tested because
of a bug in the globbing functionality of the insta crate. Globs are now
flattened and should work correctly.