Make it possible to have methods and members in tasks.
This commit makes it possible to distinguish between local variables and
members variables in tasks. Also, it makes it possible to distinguish
between functions and methods.

* A task is a struct.

* A member variable is a field inside the task-struct.

* A method is a function which takes as its first parameter a mutable
  reference to the task-struct which it is implemented for, and as second
  parameter the OperatorContext of Arcon which can be used to emit events,
  schedule timers, etc.

Rust-code for member access and method calls is represented as:

* `(self.my_member)` and `self.my_method` where `self` is the task-struct.

Parentheses are added, or else Rust will get confused in cases where the
member is a function pointer.

See

* arc-script-test/src/expect-mlir-fail-todo/task-with-funs.arc

for an example, and

* snapshots/rust@task-with-funs.arc.snap

for the Rust output