Make expect-mlir-fail-todo tests compile with arcon backend
This PR makes the expect-mlir-fail-todo tests compile. To make this
possible, the code generation of tasks has to be modified.
Arcon's `Operator`s and `OperatorContext`s are now wrapped inside a
`Handler` struct. Custom and builtin methods such as `on_event` and
`emit` can both be implemented for this struct, which makes it easier
to call methods on it, since one can just invoke `self.my_method`.

TODO:
* Rust can in some cases complain about ownership if expressions are
  nested (e.g., `self.foo(self.bar())`). Solve this by converting into
  SSA-form before generating Arcon (i.e., `let x = self.bar() in self.foo(x)`)

* Rust will complain if methods are moved into variables (since they
  have a mutable context). Checks for this need to be implemented in
  Arc-Script's ownership-checker.