Add SSA conversion for Arcon backend
This commit fixes a serious bug which would sometimes break codegen.

The following code would previously not compile:

```
task Test() ~i32 -> ~i32 {
    fun one() -> i32 {
        1
    }
    on event => emit one()
}
``

This is because the code would (in a simplified way) desugar to:

```
 0: impl Test {
 1:     fn one(&mut self) -> i32 {
 2:         1
 3:     }
 4:     fn emit(&mut self, event: i32) {
 5:         self.ctx.output(event)
 6:     }
 7:     fn handle_event(&mut self, event: i32) {
 8:         self.emit(self.one())
 9:     }
10: }
```

Rust complains that line 8 causes an ownership violation.

The solution which this commit implements is to convert into SSA form:

```
...
     fn handle_event(&mut self, event: i32) {
         let x = self.one();
         let y = self.emit(x);
         y
     }
...
```

Note that SSA form above treats `self.one` as a single expression.
It might in other words not work if we desugared to:

```
...
     fn handle_event(&mut self, event: i32) {
         let x = self;
         let y = self;
         let z = x.one();
         let w = y.emit(z);
         w
     }
...
```

Currently the MLIR and Arcon codegens use pretty much identical SSA
converters. SSA is easy work, but it might be nice to later do it at an
earlier stage (in the HIR), to avoid code-duplication.