Add support for extern functions and methods
A test which gives an idea of what is now possible to do can be viewed in:

* arc-script-test/compile/src/expect-mlir-fail-todo/extern-fun.arc
* arc-script-test/compile/src/expect-mlir-fail-todo/extern-fun.rs

One can implement functions from the outside and use them inside
arc-script. It is also possible to implement methods on the outside
for tasks defined inside arc-script.

Names of tasks are currently mangled which makes it annoying to work with them.
This will be addressed in a future patch!