Switch from trybuild to compiletest_rs
trybuild is very easy to setup, but very limited in what it can.
It is also super-slow to run since it must compile everything sequentially
with cargo.

compiletest_rs is an alternative which instead uses rustc directly. As a
matter of fact, compiletest is used within rustc itself. This commit
switches from using trybuild to compiletest_rs. One caveat is that we
need to use compiler-directives to inform compiletest_rs to output
human-readable error-messages. Also, there is now a bit more code in the
test-suite, but at least it gives more flexibility.

To accept a UI-test (test which is correctly failing to compile), you
can run `BLESS=1 cargo test`.

Additional note: If tests are failing for no reason, it means there are
duplicate build artifacts in the target/debug/deps. To solve this, do a
`cargo clean` and then `cargo test --package=arc-script-test-compile`.