Add support for ValueState
ValueState is a concept in Arcon which allows Arcon-values to be persisted
inside state-backends. An Arcon-value is a value which can be serialized
by protobuf. All values (besides streams and state) inside Arc can be
serialised as protobuf.

Value state in Arc can be used like:

    task Counter() ~i32 -> ~i32 {
        state counter: i32 = 0
        on event => {
            counter = counter + 1;
            emit counter
        }
    }

For each incoming event, the above code emits the number of times
(count) that the event has been received so far.

State requires the distinction between l-values and r-values. L-values
are values which may occur on the left-hand-side of a mutating assignment.
R-values are values which occur anywhere else.

                          counter = counter + 1
                          ^^^^^^^   ^^^^^^^
                          l-value   r-value

This desugars to a read and a write of the state variable `counter`:

    let x0 = counter.read();
    let x1 = 1;
    let x2 = x0 + x1;
    counter.write(x2)

For the moment, all state-variables have the property that they are not
values. Usage of state-variables will implicitly desugar into reads,
writes, and other expressions over values of those variables. The true
value of the state-variable is pinned and thus cannot not be moved or
aliased.

For example, the following code will return an error:

    task MapCount() ~i32 -> ~i32 {
        state map: {i32=>i32} = {};
        on event => {
            let x = map in
            x[event] = x[event] + 1;
            emit x[event]
        }
    }

While this wont:

    task MapCount() ~i32 -> ~i32 {
        state map: {i32=>i32} = {};
        on event => {
            map[event] = map[event] + 1;
            emit map[event]
        }
    }

I will add the other state types (Map, Vec, Set) in the forthcoming PRs.