Update build API and test suite
This makes it so the proc-macro API is hidden behind a feature flag,
which improves compile times considerably.

In addition all tests are updated and streamlined. Adding a test is very
easy, and it is also easy to debug it since the output is cached in the
target/out/src dir. Rust-analyzer / RLS can therefore also generate
auto-suggestions for the generated code.