Add syntax for emit-after operator
The emit after operator can be used as:

task Delay() ~i32 -> ~i32 {
    on event => emit event after 100ms
}

The associativity here is:

task Delay() ~i32 -> ~i32 {
    on event => ((emit event) after 100ms)
}

The right-hand-side operand of after must be an emit expression,
otherwise a compilation error is thrown.

The semantics of `after` are up in the air. It looks like Arcon will
support one timer at a time. If a new timer (after) is scheduled, then
previous ones are cancelled. On a high throughput stream, Delay might
therefore not emit any events at all.