Break the orphan rule for Pretty printing
As a result of the orphan rule in Rust, the Pretty struct used for
pretty printing had to be defined within the same crate as the language
construct for which the pretty printing was implemented for.

impl Display for Pretty<'i, MyLanguageConstruct, Context<'_>>

This meant all language constructs (of all six IRs) had to be defined in
the same crate. I hacked around this by inlining the pretty printer in
each module that implements the pretty printing. In other words, by
duplicating the Pretty struct I was able to break the orphan rule.