arcorn/arctime: Add unsafe impl for generated components
With this commit, all generated components can be sent between threads.
Originally this was not possible when for example a component contained
a reference counted field, since reference counted objects must not escape
to other threads, or else there will be race conditions.

After this commit we must ensure this invariant is instead upheld by the
code generator. Instead of sending the actual references, we will send
deep clones of the references.