arc: Add floating point operators
We don't currently support overloading. We can't have two functions with
the same name, but we want to be able to do arithmetics on both floats
and ints. Therefore, to support floating point operations we can
do like OCaml and abbreviate all operations with a dot such as `+.`

This commit adds support for this.