arc-lang: Add syntax-directed operators
We don't currently support overloading. We can't have two functions with
the same name, but we want to be able to do arithmetics on both floats
and ints. Therefore, to support for more kinds of numeric data types, we
can suffix the operations with the type to make it "syntax-directed".
For example

* 1.0 +f32 1.0
* 1u32 +u32 1u32
* true !=bool false

It's ugly, but there is no other good solution at the moment. Once we support
overloading we can remove this.