rust-mlir: Handle ADT generics in structs
Fix a bug in how the Rust dumper handles generic ADT types inside
structs. We have to mangle the type signature for the generic type to
something which is a valid Rust type name when creating the name for
the struct. As the type signature is no longer the same as the Rust
representation of the ADT type, we have also have to modify the MLIR
pretty printer.